refactor(front): move userStore to stores folder
